requirements.txt
Django
gunicorn
whitenoise
dj_database_url
django-allauth
